pyproject.toml
[tool.ruff]

lint.select = [
    "E",     # pycodestyle (errors)
    "W",     # pycodestyle (warnings)
    "F",     # pyflakes
    "C90",   # mccabe complexity
    "I",     # isort
    "N",     # pep8-naming
    "D",     # pydocstyle
    "UP",    # pyupgrade
    "YTT",   # flake8-2020
    "ANN",   # flake8-annotations
    "ASYNC", # flake8-async
    "S",     # flake8-bandit
    "BLE",   # flake8-blind-except
    "B",     # flake8-bugbear
    "A",     # flake8-builtins
    "C4",    # flake8-comprehensions
    "DTZ",   # flake8-datetimez
    "T10",   # flake8-debugger
    "DJ",    # flake8-django
    "EM",    # flake8-errmsg
    "EXE",   # flake8-executable
    "FA",    # flake8-future-annotations
    "ICN",   # flake8-import-conventions
    "INP",   # flake8-no-pep420
    "PIE",   # flake8-pie
    "PT",    # flake8-pytest-style
    "Q",     # flake8-quotes
    "RSE",   # flake8-raise
    "RET",   # flake8-return
    "SLF",   # flake8-self
    "SIM",   # flake8-simplify
    "TCH",   # flake8-type-checking
    "ARG",   # flake8-unused-arguments
    "PTH",   # flake8-use-pathlib
    "TD",    # flake8-todos
    "ERA",   # eradicate (commented-out code)
    "PL",    # pylint rules
    "TRY",   # tryceratops
    "FLY",   # flynt (string formatting improvements)
    "NPY",   # NumPy-specific rules
    "PERF",  # performance improvements
    "FURB",  # refurb plugin
]
lint.ignore = ["S101"]

[tool.ruff.lint.per-file-ignores]
"/leafroot/tests/*.py" = ["ARG001"]
"/leafroot/spark_apps/*.py" = ["S608"]
"/leafroot/data/alembic/*" = ["ALL"]

[tool.ruff.lint.isort]
case-sensitive = true
force-wrap-aliases = true
combine-as-imports = true

section-order = [
    "future",
    "standard-library",
    "third-party",
    "first-party",
    "local-folder",
]
# McCabe complexity settings
[tool.ruff.lint.mccabe]
max-complexity = 10

# Enable all `pydocstyle` rules, limiting to those that adhere to the
# Google convention via `convention = "google"`, below.
[tool.ruff.lint.pydocstyle]
convention = "google"


[tool.bandit]
exclude_dirs = ["tests", "scripts", "utils"]

[tool.poetry]
name = "leafroot"
version = "0.1.0"
description = "leaf root dependencies"
authors = ["mihai011 <mihaimihai011@gmail.com>"]
readme = "README.md"

[tool.pytest.ini_options]
pythonpath = "."
addopts = ["--import-mode=importlib"]

[tool.poetry.dependencies]
python = ">=3.11,<=3.14"
aiohttp = ">=3.12.15"
aioresponses = ">=0.7.8"
aiosignal = ">=1.4.0"
alembic = ">=1.16.4"
anyio = ">=3.0.0,<4"
asgiref = ">=3.9.1"
astroid = ">=3.3.11"
async-timeout = ">=5.0.1"
asyncpg = ">=0.30.0"
attrs = ">=25.3.0"
bandit = ">=1.8.6"
bcrypt = "4.0.1"                                                           # could be issues here (needs to be set to 4.0.1)
blacken-docs = ">=1.19.1"
celery = ">=5.5.3"
certifi = ">=2025.8.3"
cffi = ">=1.17.1"
charset-normalizer = ">=3.4.2"
click = ">=6.7,<8.2"
coverage = ">=7.10.2"
cryptography = ">=45.0.6"
dill = ">=0.4.0"
doc8 = ">=2.0.0"
docformatter = ">=1.7.7"
ecdsa = ">=0.19.1"
execnet = "2.1.0"
fastapi = ">=0.116.1"
fastapi-cache2 = { extras = ["redis"], version = ">=0.2.2" }
flower = ">=2.0.1"
frozenlist = ">=1.7.0"
greenlet = ">=3.2.4"
gunicorn = ">=23.0.0"
h11 = ">=0.16.0"
httpcore = ">=1.0.9"
httptools = ">=0.6.4"
httpx = ">=0.28.1"
idna = ">=3.10"
iniconfig = ">=2.1.0"
ipython = ">=9.4.0"
isort = ">=6.0.1"
jinja2 = ">=3.1.6"
lazy-object-proxy = ">=1.11.0"
maturin = ">=1.9.3"
mccabe = ">=0.7.0"
motor = ">=3.7.1"
multidict = ">=6.6.3"
mypy = ">=1.17.1"
mypy-extensions = ">=1.1.0"
orjson = ">=3.11.1"
packaging = ">=25.0"
passlib = ">=1.7.4"
pathspec = ">=0.12.1"
pika = ">=1.3.2"
platformdirs = ">=4.3.8"
pluggy = ">=1.6.0"
pre-commit = ">=4.2.0"
psycopg2 = ">=2.9.10"
py = ">=1.11.0"
pyasn1 = ">=0.6.1"
pycparser = ">=2.22"
pydantic = { extras = ["email"], version = ">=2.11.7" }
pyparsing = ">=3.2.3"
pytest = ">=8.4.1"
pytest-asyncio = ">=1.1.0"
pytest-cov = ">=6.2.1"
pytest-cover = ">=3.0.0"
pytest-coverage = ">=0.0"
pytest-forked = ">=1.6.0"
pytest-xdist = "3.8.0"
python-dotenv = ">=1.1.1"
python-jose = ">=3.5.0"
python-multipart = ">=0.0.20"
pyyaml = ">=6.0.2"
redis = ">=4.2.0rc1,<5.0.0"
requests = ">=2.32.4"
rfc3986 = ">=2.0.0"
rsa = ">=4.9.1"
six = ">=1.17.0"
sniffio = ">=1.3.1"
sqlalchemy = "2.0.28"
sqlalchemy-serializer = "1.4.12"
sqlalchemy-utils = ">=0.41.2"
starlette = ">=0.47.2"
starlette-context = ">=0.4.0"
toml = ">=0.10.2"
tomli = ">=2.2.1"
tomlkit = ">=0.13.3"
tqdm = ">=4.67.1"
typing-extensions = ">=4.14.1"
uvicorn = ">=0.35.0"
uvloop = ">=0.21.0"
watchgod = ">=0.8.2"
websockets = ">=15.0.1"
wrapt = ">=1.17.2"
yarl = ">=1.20.1"
sentry-sdk = ">=2.34.1"
polars = ">=1.32.2"
pyarrow = ">=21.0.0"
scylla-driver = ">=3.29.3"
surrealdb = ">=1.0.6"
faker = ">=37.5.3"
pyspark = "3.5.0"
pandas = ">=2.3.1"
grpcio = ">=1.74.0"
google = ">=3.0.0"
google-api-python-client = ">=2.178.0"
grpcio-status = ">=1.74.0"
timeout-decorator = ">=0.5.0"
kafka-python = ">=2.2.15"
json-log-formatter = ">=1.1.1"
faust-streaming = { git = "https://github.com/faust-streaming/faust.git" }
python-logstash-async = ">=4.0.2"
pip = ">=25.2"
nest-asyncio = ">=1.6.0"
aioboto3 = ">=15.0.0"
minio = ">=7.2.16"
miniopy-async = "1.11"
pillow = ">=11.3.0"
diagrams = ">=0.24.4"
pytest-timeout = ">=2.4.0"
graphviz = ">=0.13.2,<0.21.0"
aiokafka = ">=0.12.0"
pydantic-settings = "^2.10.1"
pyiceberg = ">=0.9.1"
psycopg2-binary = "^2.9.10"
ruff = "^0.12.8"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
